# contracts/openapi/root.yaml
openapi: 3.1.0
info:
  title: QuantPilot Runtime API
  version: 2.3.3
  summary: QuantPilot v2.3.3 HTTP 控制面契约
  description: |
    QuantPilot 策略开发与运行平台的 HTTP API 契约。
    覆盖 capability discovery、runtime session、compile、backtest、AI 提案与审批、
    沙箱验证、告警、快照、runbook、混沌实验、运营报表等全部控制面端点。

    API 版本前缀说明:
      /api/    — v0.1 遗留, 覆盖编译/图管理/运行时/热插拔等核心策略开发循环
      /api/v1/ — 较新系统端点 (告警/快照/审批/混沌/runbook/报表/存储)
    此为有意设计, 见各端点的 x-chain-stage 标注。
servers:
  - url: http://127.0.0.1:3000
    description: 本地开发服务器
...
paths:
...
  /api/v1/snapshots/{snapshot_id}/restore:
    post:
      operationId: restoreSnapshot
      summary: 一键恢复到指定快照
      tags: [snapshots]
      x-owner: backend-runtime
      x-chain-stage: system
      x-stability: beta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [actor_id]
              properties:
                actor_id:
                  type: string
                reason:
                  type: string

  # ── Runbook ──
...